LPS-132023 Move getSLAStatusIconInfo to utils and adjust its return
